Update RailroadPluginAPI version and improve heartbeat logic using new features.
Bumped RailroadPluginAPI dependency to 1.0.3. Refactored WakatimePlugin to use new services, improved heartbeat event handling to set line number, cursor position, project, and language, and fixed unsaved file logic for file events.

NOTE: The current version of railroad auto saves every time the file is modified. This means that two heartbeats are sent every time the file is changed. In the future, Railroad will batch saving so this wont be an issue (as well as allow the user to disable auto saving).
NOTE: There is currently a bug with the file numbers in Railroad. This bug introduces ghost lines that you cant navigate to. Due to how the document service works, these are sent in the Wakatime heartbeat, meaning the number of lines is often higher than the number of navigatable lines.